Do not wait for SSH keys if there is an external controller
After stop reconciling KubevirtCluster with external controller, the machine reconcile function is waiting for an SSH keys that are not set anymore, because the cluster reconcile is never called.

To release this block, this PR does two things:

First, it add condition to the machine reconcile, to not wait for SSH keys if there is an external controller.

Then, the implementation of the `KubevirtMachineReconciler.reconcileDelete` function requires an SSH key to build a new machine instance, while the key is not actually used in the deletion process. To ease that, this PR introduces a new version of the `NewMachine` function: the `NewBasicMachine` function. This function creates a new machine instance only with the fields needed for deletion. Then this new function is used in the `NewMachine` function to populate the common fileds, and in the `KubevirtMachineReconciler.reconcileDelete` function to return an instance with fields needed in the deletion process.

Signed-off-by: Alex Gradouski <agradouski@apple.com>