Make secrets for third-party PRs available
If a PR gets labeled with `ok-to-test` the integration test job is
triggered and this github action has access to secrets.

The github action does for security resaons not allow secrets on all
PRs, so we add this extra step.

Still, the action itself, while using a secret, does not have any
interaction points with users. The PR content is actually not even
checked out.

Signed-off-by: Roman Mohr <rmohr@redhat.com>