Fix bug in KubevirtCluster Reconciler - endless reconcile loop with external controller
When the KubevirtCluster CR is manage by an external controller, sometimes there is an update race between capk and the external controller, where each update in one hand, triggers reconcile and update in the other hand and vice versa.

This PR changes that so in this case, the CR is filtered out from the reconciling watch, and so capk will ignore changes made by an external controller.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>