Run goimports as part of the sanity test
run goimports to sort the imports in go files. goimports also performs gofmt.

To make it work, this commit also contains the sorted files as well.

This PR require a change in the way we're used to push PRs: it requires to make sure that the format of the files is valid before pushing the commits. For that, the PR also contain additions to the CONTRIBUTING.md file with the relevant instructions.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>