Graceful deletion of a VirtualMachineInstance
When a host node is deleted, the guest VMIs are evicted, killing the
running tenant node in this VMI.

This PR gracefully delete the VMI, by first drain the guest node. Only
if the drain successful, CAPK will delete the VMI.

The PR is based on a feature in KubeVirt, where by setting the eviction
strategy to "external" KubeVirt will not delete the VMI, but will set
the `vmi.Status.EvacuationNodeName` field to signal to the external
controller (CAPK in this case) that the VMI should be evicted.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>