e2e, conformance: introduce conformance tests
Conformance tests are executed with sonobuoy, which
run from a pod deployed in infra cluster.

The conformance e2e tests themselve are running in
the tenant cluster.
The pod mounts the tenant cluster kubeconfig
from the <cluster-name>-kubeconfig secret.

The run-conformance.sh script creates a configMap
with names of tests that need to be skipped due to
issues with DNS (accessing services via internal
domain name) in the guest cluster.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>