Option to poll externally managed VMs on a recurring interval
Externally manged VMs, meaning VMs which are not on the same local cluster
that capk is running on, are not watched by informers. Instead we need to
periodically wake up the reconcile loop to detect changes to these VMs.
This PR adds an option that allows setting the interval the
kubevirtMachine controller wakes up at for externally manged VMs

Signed-off-by: David Vossel <davidvossel@gmail.com>