Remove the IMAGE_REPO env var from all the temlates
cluster-api project wants to stop using the k8s.gcr.io repository, as
it's too expensive. After checking in other cluster API project, it
seems that this value is no longer in the template.

This PR removes all the usage in IMAGE_REPO variable; for that, the PR
also upgrade the clusterctl in the kubevirtci script, to v1.2.4

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>