Remove the setup counter from the KVMachine ready status
Cluster-API takes the reason and the message from the infra-machine (in
our case, the KubeVirtMachine) `Ready` condition, into `Machine`
`InfrastructureReady` condition. Currently the text is something like
`"0 of 2 completed"`. This is not very informative.

This PR changes the behavior so now the KV Machine Ready condition will
include both the reason and the message from the failed condition.

That way, the cluster-api `InfrastructureReady` condition, will also
contain the same reason and message. This will make the actual issue, if
exists, to more accessible to the user.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>