Allow annotate and label KubevirtClusterTemplates
The current CRD does not allow modification of any field in KubevirtClusterTemplate
resources. That prevent user to add annotations or labels to the
templates, and that does not follow the regular behavior of K8s.

In addition, the new capi is trying to annotate templates, and failed to
do that, and this is an actual bug.

This PR fixes the issue by making only the `Spec` field as read only,
letting modifiying the metadata fields as in any K8s resource.