Fix the e2e test
The e2e test constantly faling, when it check that the clusterctl output
to stderr is empty, while clusterctl prints a warning (not a real
error): `"version available: v1.9.3 -> v1.9.4"`.

This asserion is worng, as we don't test clusterctl.

Also, in one case, the test assertion error message prints the stdout
instead of stderr, making it harder to debug.

This PR fixes the issue by removing the clusterctl stderr empty assertion,
replacing it with simple log, if stderr is not empty, and by downloading
the latest clusterctl.

In addition, this PR fixes the wrong printout.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>